Simplified data flow and code redundancy by using a single React page for getting all categories under general aptitude
