+++ src/App.js
@@ -4,8 +4,8 @@ import AdminHome from './Pages/AdminHome';
 import Login from './Pages/Login';
 import LandingPage from './Pages/LandingPage';
 import PoTFormulas from './Subtopics/PoTFormulas';
-import PoTQuestions from './Subtopics/PoTQuestions';
 import SubtopicsHome from './Pages/SubtopicsHome';
+import GeneralQuestions from './Subtopics/GeneralQuestions';
 
 function App() {
   return (
@@ -17,9 +17,7 @@ function App() {
         <Route path='/' element={<LandingPage/>}/>
         <Route path='/subtopicshome' element={<SubtopicsHome/>}/>
         <Route path='/potformulas' element={<PoTFormulas/>}/>
-        <Route path='/potquestions' element={<PoTQuestions/>}/>
-        
-        
+        <Route path='/generalquestions' element={<GeneralQuestions/>}/>
       </Routes>
 
    </>
