+++ src/Pages/SubtopicsHome.jsx
@@ -5,14 +5,21 @@ function SubtopicsHome() {
 // calling the state we created in the link using the hook useLocation 
   const location = useLocation()
   const passedData = location.state?.data;
-
-
-
+ 
   return (
     <div className='d-flex flex-column'>
       {/* here we are calling the data passed in the link so we dont need to type the route and name of the link manually in this page  */}
-    <Link to={passedData.route1}>{passedData.route1Name}</Link>
-    <Link to={passedData.route2}>{passedData.route2Name}</Link>
+    {passedData.length>0?
+    passedData.map(item=>(
+      <Link to={item.route} state={{data:{
+        section_name:item.section_name,
+        category:item.category
+      }
+    }}>{item.routeName}</Link>
+    ))
+      :
+      <p>Nothing to display!</p>
+    }
   </div>
   )
 }
